third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFEPointLightElement-svgdom-x-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'x' property of the SVGFEPointLightElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS pointLight.x.baseVal is 50
PASS pointLight.x.baseVal is 100
PASS successfullyParsed is true

TEST COMPLETE
